refactor(openai): move the Chat Completions adapter to openai/completions
The Chat Completions adapter moves from genai/openai to
genai/openai/completions, and its package is renamed accordingly.
genai/openai becomes the home for every OpenAI wire protocol this
library implements: a Responses API adapter joins the family, and the
two protocols live as sibling packages so the import paths say which
wire a Model speaks.

No compatibility layer for the old import path: the examples, the
README and the .agents docs move in the same change. The README gains a
provider-dialects section covering the nil default, the three shipped
dialects and the replay-shape knob, and .agents records the move.
