+++ examples/openai-client/main.go
@@ -28,7 +28,7 @@ import (
 	"google.golang.org/adk/v2/session"
 	"google.golang.org/genai"
 
-	genaiopenai "github.com/achetronic/adk-utils-go/genai/openai"
+	"github.com/achetronic/adk-utils-go/genai/openai/completions"
 )
 
 func main() {
@@ -36,7 +36,7 @@ func main() {
 
 	// 1. Create the OpenAI client
 	//    This is all you need to switch from Gemini to OpenAI
-	llmModel := genaiopenai.New(genaiopenai.Config{
+	llmModel := completions.New(completions.Config{
 		APIKey:    os.Getenv("OPENAI_API_KEY"),
 		BaseURL:   getEnvOrDefault("OPENAI_BASE_URL", ""),
 		ModelName: getEnvOrDefault("MODEL_NAME", "gpt-4o"),
