+++ examples/session-memory/main.go
@@ -32,7 +32,7 @@ import (
 	"google.golang.org/adk/v2/tool"
 	"google.golang.org/genai"
 
-	genaiopenai "github.com/achetronic/adk-utils-go/genai/openai"
+	"github.com/achetronic/adk-utils-go/genai/openai/completions"
 	sessionredis "github.com/achetronic/adk-utils-go/session/redis"
 )
 
@@ -166,8 +166,8 @@ func runAgent(ctx context.Context, runnr *runner.Runner, sessionID string, input
 	return responseText
 }
 
-func getOpenAIModel() *genaiopenai.Model {
-	return genaiopenai.New(genaiopenai.Config{
+func getOpenAIModel() *completions.Model {
+	return completions.New(completions.Config{
 		APIKey:    os.Getenv("OPENAI_API_KEY"),
 		BaseURL:   getEnvOrDefault("OPENAI_BASE_URL", "http://localhost:11434/v1"),
 		ModelName: getEnvOrDefault("MODEL_NAME", "qwen3:8b"),
