+++ genai/openai/completions/dialect_deepseek.go
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2026 Alby Hernández <hola@achetronic.com>
 // SPDX-License-Identifier: Apache-2.0
 
-package openai
+package completions
 
 // DeepSeek is the dialect for DeepSeek: reasoning travels as the plain-text
 // field, and the provider requires it back on assistant turns.
