+++ genai/openai/completions/wire_test.go
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2026 Alby Hernández <hola@achetronic.com>
 // SPDX-License-Identifier: Apache-2.0
 
-package openai
+package completions
 
 import (
 	"context"
